Added holders and signatures

+++ api/app/models/holder.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 from datetime import datetime
-from sqlalchemy import Column, Integer, PrimaryKeyConstraint, String, ForeignKey, TIMESTAMP
+from sqlalchemy import Column, Integer, PrimaryKeyConstraint, String, ForeignKey, TIMESTAMP, BigInteger
 from sqlalchemy.orm import relationship
 from app import Base
 
@@ -9,8 +9,8 @@ class Holder(Base):
     __tablename__ = "holder"
     address = Column(String, nullable=False)
     token_id = Column(Integer, ForeignKey("token.id"), nullable=False)
-    initial_balance = Column(Integer, nullable=False)
-    current_balance = Column(Integer, nullable=False)
+    initial_balance = Column(BigInteger, nullable=False)
+    current_balance = Column(BigInteger, nullable=False)
     last_checked = Column(TIMESTAMP, nullable=False)
     __table_args__ = (PrimaryKeyConstraint("address", "token_id"),)
     token = relationship("Token", back_populates="holders")
